Fix minor test issue
Fixed a minor issue causing a test to fail.

The code is able to build and run. May need some extra test cases.

While the function itself works just fine,
not sure if the implementation is done correctly.

One last commit needed to check if all changes are up
to coding standards.